Update install.sh
Updated for new commits

added the -f (force) option. Helps whenever there's corruption or permissions issue.

wget will now resume existing downloads. Untested. Questions.
If the file is corrupted or different from what's expected, will it overwrite it?
If the file already exists and is completed, will it skip it?


Untested behavior in the case the file can't be resumed, I don't know if the rm -f is needed now. It might be a good idea to not delete it at all in case the install fails at a later part of the process. The user can resume the install without having to redownload the file if you don't delete it. Files in /tmp/ get removed by themselves after shutting down the device anyways.