Create and Use SSL for SillyTavern
Added ability for launcher to generate and implement SSL for SillyTavern.

Auto-creates SSL with OpenSSL bundled with GIT.
After creation updates ST launch links to use the cert.
Added abilities to delete and regen cert.
Added links to read me on why self-signed certs are safe in this case and how to add to trusted.