Update Launcher.bat
I learned more about how batch works, so I updated the GPU detection part to pick the best from among all the GPUs instead of the last two. Also altered how the script stores the GPU information to the gpu_info log file, and how it fetches the GPU info to display it on the home screen.
Made it so that there is no ANSI code being sent with our GPU info, which makes it much nicer to deal with.
Less ANSI codes in our data, the better.
ANSI is only good for on screens, not returned data; and it looks ugly in log files.

I have tested this by installing a new text-completion model and it works just fine. I only have two GPUs though, so someone with one or three would need to test it as well.