Replace gpu_info.bat with gpu_info.ps1
A better version of gpu_info.bat.

This script collects the name and vram information from all installed GPUs, puts it into a table, sorts that table from highest-lowest based on the VRAM entry, then picks the first GPU (which will have the most VRAM).

It then sends that data out without any fancy formatting, so we can parse it nicely. We can add colors when we are displaying the data this sends. ANSI code should not be in data we intend to manipulate.